+++ cg/meta/workflow/balsamic.py
@@ -339,7 +339,7 @@ class BalsamicAnalysisAPI:
             )
         LOG.info("")
 
-    def get_sample_params(self, case_id: str, panel_bed: str) -> dict:
+    def get_sample_params(self, case_id: str, panel_bed: Optional[str]) -> dict:
 
         """Returns a dictionary of attributes for each sample in given family,
         where SAMPLE ID is used as key"""
@@ -355,7 +355,17 @@ class BalsamicAnalysisAPI:
         self.print_sample_params(case_id=case_id, sample_data=sample_data)
         return sample_data
 
-    def resolve_target_bed(self, panel_bed, link_object: models.FamilySample) -> Optional[str]:
+    def check_application_type_wes(self, case_id: str) -> bool:
+        """Checks if any of the samples in case have application type WES"""
+        for link_object in self.get_balsamic_sample_objects(case_id=case_id):
+            sample_application_type = self.get_application_type(link_object)
+            if str(sample_application_type).lower() == "wes":
+                return True
+        return False
+
+    def resolve_target_bed(
+        self, panel_bed: Optional[str], link_object: models.FamilySample
+    ) -> Optional[str]:
         if panel_bed:
             return panel_bed
         if self.get_application_type(link_object) not in self.__BALSAMIC_BED_APPLICATIONS: