+++ tests/conftest.py
@@ -233,6 +233,19 @@ def fixture_orderform(fixtures_dir: Path) -> Path:
     return _path
 
 
+@pytest.fixture(name="mip_dna_store_files")
+def fixture_mip_dna_store_files(apps_dir: Path) -> Path:
+    """Return the path to the directory with mip dna store files"""
+    _path = apps_dir / "mip" / "dna" / "store"
+    return _path
+
+
+@pytest.fixture(name="mip_deliverables_file")
+def fixture_mip_deliverables_files(mip_dna_store_files: Path) -> Path:
+    """Fixture for general deliverables file in mip"""
+    return mip_dna_store_files / "case_id_deliverables.yaml"
+
+
 # Orderform fixtures