+++ CHANGELOG.md
@@ -13,6 +13,25 @@ Try to use the following format:
 ### Changed
 ### Fixed
 
+
+## [13.12.0]
+### Added
+- Store all available completed microbial analyses in HK
+
+## [13.10.1]
+
+### Fixed
+- Storing chromograph, upd and rhocall files in housekeeper
+
+## [13.11.0]
+
+### Changed
+- Balsamic always skips mutect when application is WES
+- SPRING compression is set to run on oldest families first
+
+### Fixed
+- Format print statements
+
 ## [13.10.1]
 
 ### Fixed
@@ -22,6 +41,9 @@ Try to use the following format:
 
 ### Added
 - Functionality to deploy `genotype` with CG on hasta
+
+### Fixed
+- Stored completed not parsing through all completed entries
 
 ## [13.9]